+++ CHANGELOG.md
@@ -3,3 +3,7 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
 - Added the `itemToPatternMappings` method to `DecoratedPotPatternRegistryEntrypoint`.
   - Removed `SherdRegistry`.
+- Added the `frozenlib:scares_piglin` Entity Type tag.
+- Added the `frozenlib:blazes` Entity Type tag.
+- Added the `frozenlib:hoglins` Entity Type tag.
+- Added the `frozenlib:ghost_like` Entity Type tag.